docs: Migrate publishing to 'builds'
Instead of publishing to gh-pages, which supports a single set of pages,
publish the results to the S3 repo for builds.zephyrproject.org, with
PRs getting their own tree, and the 'main' getting replaced each time
main merges.

The 'zephyr' repo, when it incorporates the docs, will generate them
itself and publish them as part of the normal doc updates for 'latest'
and for releases.

Signed-off-by: David Brown <david.brown@linaro.org>