samples: philosophers: Add statistics
Using the newly provided `sync::Mutex`, implement support for printing
statistics about the philosophers habits.  Remove the verbose message
showing each meal, and replace with a period print of the statics.

Change the sample.yml file with a regexp to make sure that all of the
philosphers are able to eat at least a bit.

Signed-off-by: David Brown <david.brown@linaro.org>