rust: Add assertions, overflow checks and debug
With a previous cmake change, Rust builds are now done in `release` mode
by default.  Add profile directives to all of the samples and tests to
ensure overflow checks, and debug assertions are enabled, as well as to
still emit debugging symbols.  This more closely matches the default
configuration of the Zephyr code.

General advice would be to keep the overflow checks, but disabling the
debug checks might be useful after sufficient testing.

Because debug symbols are not included in the target image, there is
little reason to not always include them.

The structure of cargo requires these flags to be set in each
application, with the defaults being builtin.

Signed-off-by: David Brown <david.brown@linaro.org>