cmake: Add ZEPHYR_BASE to environment in template
Export the ZEPHYR_BASE environment variable to the generated
`.cargo/config.toml` file.  The regular zephyr build system will set
this if it is not set, but then trying to run a standalone build will
cause the bindgen build to fail, as it needs this value.  Explicitly
place it into the environment to allow this to work.

Signed-off-by: David Brown <david.brown@linaro.org>